Rename Java builder and replace Java UI activator
