+++ bundles/tools.vitruv.domains.java.ui/META-INF/MANIFEST.MF
@@ -4,7 +4,7 @@ Bundle-Name: Vitruv Domain Java UI
 Bundle-SymbolicName: tools.vitruv.domains.java.ui;singleton:=true
 Automatic-Module-Name: tools.vitruv.domains.java.ui
 Bundle-Version: 1.1.0.qualifier
-Bundle-Activator: tools.vitruv.domains.java.ui.monitorededitor.activator.Activator
+Bundle-Activator: tools.vitruv.domains.java.ui.Activator
 Require-Bundle: org.eclipse.xtend.lib,
  org.eclipse.ui,
  org.eclipse.core.runtime,
@@ -19,9 +19,9 @@ Require-Bundle: org.eclipse.xtend.lib,
  edu.kit.ipd.sdq.activextendannotations
 Bundle-RequiredExecutionEnvironment: JavaSE-11
 Bundle-ActivationPolicy: lazy
-Export-Package: tools.vitruv.domains.java.ui.monitorededitor,
+Export-Package: tools.vitruv.domains.java.ui;x-internal:=true,
+ tools.vitruv.domains.java.ui.monitorededitor,
  tools.vitruv.domains.java.ui.builder,
- tools.vitruv.domains.java.ui.monitorededitor.activator;x-internal:=true,
  tools.vitruv.domains.java.ui.monitorededitor.astchangelistener,
  tools.vitruv.domains.java.ui.monitorededitor.astchangelistener.classification,
  tools.vitruv.domains.java.ui.monitorededitor.astchangelistener.classification.postchange;x-internal:=true,
